Refactor dashboard JS into modular front-end files
