+++ src/singular/dashboard/templates/dashboard.html
@@ -345,7 +345,7 @@
   <pre id='live-events' class='live-events-box'></pre>
 </section>
 
-<script src='/static/dashboard.js'></script>
+<script type='module' src='/static/dashboard.js'></script>
 
 </body>
 </html>
